Rename browser policy files
