+++ sdk/storage/storage-blob/src/policies/StorageBrowserPolicy.ts
@@ -14,25 +14,25 @@ import { HeaderConstants, URLConstants } from "../utils/constants";
 import { setURLParameter } from "../utils/utils.common";
 
 /**
- * BrowserPolicy will handle differences between Node.js and browser runtime, including:
+ * StorageBrowserPolicy will handle differences between Node.js and browser runtime, including:
  *
  * 1. Browsers cache GET/HEAD requests by adding conditional headers such as 'IF_MODIFIED_SINCE'.
- * BrowserPolicy is a policy used to add a timestamp query to GET/HEAD request URL
+ * StorageBrowserPolicy is a policy used to add a timestamp query to GET/HEAD request URL
  * thus avoid the browser cache.
  *
  * 2. Remove cookie header for security
  *
  * 3. Remove content-length header to avoid browsers warning
  *
- * @class BrowserPolicy
+ * @class StorageBrowserPolicy
  * @extends {BaseRequestPolicy}
  */
-export class BrowserPolicy extends BaseRequestPolicy {
+export class StorageBrowserPolicy extends BaseRequestPolicy {
   /**
-   * Creates an instance of BrowserPolicy.
+   * Creates an instance of StorageBrowserPolicy.
    * @param {RequestPolicy} nextPolicy
    * @param {RequestPolicyOptions} options
-   * @memberof BrowserPolicy
+   * @memberof StorageBrowserPolicy
    */
   constructor(nextPolicy: RequestPolicy, options: RequestPolicyOptions) {
     super(nextPolicy, options);
@@ -43,7 +43,7 @@ export class BrowserPolicy extends BaseRequestPolicy {
    *
    * @param {WebResource} request
    * @returns {Promise<HttpOperationResponse>}
-   * @memberof BrowserPolicy
+   * @memberof StorageBrowserPolicy
    */
   public async sendRequest(request: WebResource): Promise<HttpOperationResponse> {
     if (isNode) {
