+++ sdk/storage/storage-file-share/src/index.browser.ts
@@ -24,7 +24,7 @@ export * from "./policies/CredentialPolicy";
 export * from "./RetryPolicyFactory";
 export * from "./TelemetryPolicyFactory";
 export * from "./UniqueRequestIDPolicyFactory";
-export * from "./BrowserPolicyFactory";
+export * from "./StorageBrowserPolicyFactory";
 export * from "./ShareServiceClient";
 export { CommonOptions } from "./StorageClient";
 export * from "./generatedModels";
